forward developing

+++ adjointflows/kernel/model_setup.py
@@ -1,29 +1,32 @@
-from tools.job_utils import remove_file, clean_symlink_target, wait_for_launching
-
+from tools.job_utils import remove_file, clean_symlink_target, wait_for_launching, check_path_is_correct
+from tools import GLOBAL_PARAMS
 import subprocess
 import logging
 import os
 
 class ModelGenerator:
     
-    def __init__(self, base_dir, current_model_num):
+    def __init__(self, current_model_num):
         """
         Args:
             base_dir (str): The work directorty for model generation (it shouold be specfem3d/)
             current_model_num (int): The current model number
         """
-        self.base_dir          = base_dir
+        self.base_dir          = GLOBAL_PARAMS['base_dir']
         self.current_model_num = current_model_num
         self.specfem_dir       = os.path.join(self.base_dir, 'specfem3d')
         self.databases_mpi_dir = os.path.join(self.specfem_dir, 'DATABASES_MPI')
         self.model_ready_file  = os.path.join(self.specfem_dir, 'OUTPUT_FILES', 'model_database_ready') 
     
-    
     def model_setup(self, mesh_flag=True):
         """
         Setup the model mesh and model generation
         """
-        os.chdir(self.specfem_dir)
+        if not check_path_is_correct(self.specfem_dir):
+            error_message = f"STOP: the current directory is not {self.specfem_dir}!"
+            logging.error(error_message)
+            raise ValueError(error_message)
+        
         remove_file(self.model_ready_file)
         if mesh_flag:
             self.model_mesh_generation()
@@ -51,3 +54,4 @@ class ModelGenerator:
 
         wait_for_launching(check_file = 'OUTPUT_FILES/model_database_ready', 
                            message = "model databases have been generated\n")
+
